refactor(domain, ui): 당첨 등수별 횟수 자료형 int 에서 long 으로 변환

+++ src/main/java/lotto/ui/OutputPrinter.java
@@ -26,7 +26,7 @@ public class OutputPrinter {
             WinningNumbers key = allWinningNumbers[winningNumberIndex];
             int matches = key.getMatch();
             long reward = key.getReward();
-            int winningCounter = winningCase.get(key);
+            long winningCounter = winningCase.get(key);
             DecimalFormat commaInMoney = new DecimalFormat("###,###");
 
             System.out.println(matches + "개 일치 (" + commaInMoney.format(reward) + "원) - " + winningCounter + "개");
